requirements.txt
fastapi
uvicorn[standard]
google-generativeai
pillow
python-dotenv
